# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ricon Technology Career & Learning Ecosystem API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Register a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get a student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all courses
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [ongoing, upcoming, all]
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a new course (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a course by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "404":
          description: Not found
    patch:
      operationId: updateCourse
      tags: [courses]
      summary: Update a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
...
  /internships:
    get:
      operationId: listInternships
      tags: [internships]
      summary: List all internship projects
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of internships
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Internship"
    post:
      operationId: createInternship
      tags: [internships]
      summary: Create a new internship project (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InternshipInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Internship"
  /internships/{id}:
    get:
      operationId: getInternship
      tags: [internships]
      summary: Get an internship by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Internship
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Internship"
        "404":
          description: Not found
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List all internship applications
      parameters:
        - name: studentId
          in: query
          schema:
            type: integer
        - name: internshipId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Submit an internship application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get an application by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update application status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
...
  /roadmaps/{id}:
    get:
      operationId: getRoadmap
      tags: [roadmaps]
      summary: Get a roadmap by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Roadmap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Roadmap"
        "404":
          description: Not found
...
components:
  schemas:
...
    Student:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        college:
          type: string
        degree:
          type: string
        graduationYear:
          type: integer
        skills:
          type: ["string", "null"]
        interests:
          type: ["string", "null"]
        role:
          type: string
          enum: [student, mentor, admin]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - fullName
        - email
        - phone
        - college
        - degree
        - graduationYear
        - role
        - createdAt
    StudentInput:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        college:
          type: string
        degree:
          type: string
        graduationYear:
          type: integer
        skills:
          type: string
        interests:
          type: string
        role:
          type: string
          enum: [student, mentor, admin]
        avatarUrl:
          type: string
      required:
        - fullName
        - email
        - phone
        - college
        - degree
        - graduationYear
    StudentUpdate:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        college:
          type: string
        degree:
          type: string
        graduationYear:
          type: integer
        skills:
          type: string
        interests:
          type: string
        avatarUrl:
          type: string
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [ongoing, upcoming]
        duration:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        enrollmentCount:
          type: integer
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - category
        - status
        - duration
        - level
        - enrollmentCount
        - price
        - createdAt
    CourseInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [ongoing, upcoming]
        duration:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        price:
          type: number
        imageUrl:
          type: string
        tags:
          type: string
      required:
        - title
        - description
        - category
        - status
        - duration
        - level
        - price
    CourseUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [ongoing, upcoming]
        duration:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        price:
          type: number
        imageUrl:
          type: string
        tags:
          type: string
...
    Internship:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        duration:
          type: string
        stipend:
          type: ["string", "null"]
        skills:
          type: string
        mentorName:
          type: ["string", "null"]
        openings:
          type: integer
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - category
        - duration
        - skills
        - openings
        - createdAt
    InternshipInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        duration:
          type: string
        stipend:
          type: string
        skills:
          type: string
        mentorName:
          type: string
        openings:
          type: integer
        imageUrl:
          type: string
      required:
        - title
        - description
        - category
        - duration
        - skills
        - openings
    Application:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        internshipId:
          type: integer
        status:
          type: string
          enum: [applied, approved, allocated, reviewing, completed, rejected]
        appliedAt:
          type: string
        studentName:
          type: ["string", "null"]
        internshipTitle:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - id
        - studentId
        - internshipId
        - status
        - appliedAt
    ApplicationInput:
      type: object
      properties:
        studentId:
          type: integer
        internshipId:
          type: integer
        notes:
          type: string
      required:
        - studentId
        - internshipId
    ApplicationUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [applied, approved, allocated, reviewing, completed, rejected]
        notes:
          type: string
...
    Roadmap:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [mtech, fullstack, aiml, dsa, placement]
        steps:
          type: string
        icon:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
      required:
        - id
        - title
        - description
        - type
        - steps
...