# lib/api-zod/src/generated/types/internship.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ricon Technology Career & Learning Ecosystem API
 * OpenAPI spec version: 0.1.0
 */

export interface Internship {
  id: number;
  title: string;
  description: string;
  category: string;
  duration: string;
  /** @nullable */
  stipend?: string | null;
  skills: string;
  /** @nullable */
  mentorName?: string | null;
  openings: number;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}