# lib/api-zod/src/generated/types/internshipInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ricon Technology Career & Learning Ecosystem API
 * OpenAPI spec version: 0.1.0
 */

export interface InternshipInput {
  title: string;
  description: string;
  category: string;
  duration: string;
  stipend?: string;
  skills: string;
  mentorName?: string;
  openings: number;
  imageUrl?: string;
}