# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ricon Technology Career & Learning Ecosystem API
 * OpenAPI spec version: 0.1.0
 */
import type { StudentRole } from './studentRole';

export interface Student {
  id: number;
  fullName: string;
  email: string;
  phone: string;
  college: string;
  degree: string;
  graduationYear: number;
  /** @nullable */
  skills?: string | null;
  /** @nullable */
  interests?: string | null;
  role: StudentRole;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}