# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ricon Technology Career & Learning Ecosystem API
 * OpenAPI spec version: 0.1.0
 */
import type { StudentInputRole } from './studentInputRole';

export interface StudentInput {
  fullName: string;
  email: string;
  phone: string;
  college: string;
  degree: string;
  graduationYear: number;
  skills?: string;
  interests?: string;
  role?: StudentInputRole;
  avatarUrl?: string;
}